Bootstrap-function-declarations
