+++ src/Symfony/Component/DependencyInjection/composer.json
@@ -39,6 +39,9 @@
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\DependencyInjection\\": "" },
+        "files": [
+            "Loader/Configurator/functions.php"
+        ],
         "exclude-from-classmap": [
             "/Tests/"
         ]
